rag/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
python-multipart
pydantic